Lift MouseButtons up one directory as part of a refactoring of the code that uses it
